float change to double readerbug??

+++ branches/D4D/JANE.Shop/Persistence/JaneShopOrdersDao.cs
@@ -75,7 +75,7 @@ namespace JANE.Shop.Persistence
                          m.Address = record.GetStringOrEmpty(4);
                          m.Email = record.GetStringOrEmpty(5);
                          m.Mobile = record.GetStringOrEmpty(6);
-                         m.Paymoney = record.GetFloat(7);
+                         m.Paymoney = record.GetDouble(7);
                          m.Paytype =(PayType)( record.GetInt32OrDefault(8,0));
                          m.Payresult = (PayResult)(record.GetInt32OrDefault(9, 0));
                          m.Payremark = record.GetStringOrEmpty(10);
@@ -99,7 +99,7 @@ namespace JANE.Shop.Persistence
             m.Address = record.GetStringOrEmpty(4);
             m.Email = record.GetStringOrEmpty(5);
             m.Mobile = record.GetStringOrEmpty(6);
-            m.Paymoney = record.GetFloat(7);
+            m.Paymoney = record.GetDouble(7);
             m.Paytype = (PayType)(record.GetInt32OrDefault(8, 0));
             m.Payresult = (PayResult)(record.GetInt32OrDefault(9, 0));
             m.Payremark = record.GetStringOrEmpty(10);
